Add Condomínio ook

+++ Projeto_TCC/Model/Funcionarios.cs
@@ -9,7 +9,7 @@ namespace Projeto_TCC.Model
     class Funcionarios
     {
         private string func_Nome;
-        private int func_Cpf;
+        private long func_Cpf;
         private Condominios condominios;
         private string func_Funcao;
         private string func_Telefone;
@@ -22,7 +22,7 @@ namespace Projeto_TCC.Model
         }
 
         public string Func_Nome { get => func_Nome; set => func_Nome = value; }
-        public int Func_Cpf { get => func_Cpf; set => func_Cpf = value; }
+        public long Func_Cpf { get => func_Cpf; set => func_Cpf = value; }
         public string Func_Funcao { get => func_Funcao; set => func_Funcao = value; }
         public string Func_Telefone { get => func_Telefone; set => func_Telefone = value; }
         public string Func_Celular { get => func_Celular; set => func_Celular = value; }
